llscan: Add nodeinfo command
Add the nodeinfo command to display basic information about the process
taken from the Node.js global process object.
This command requires a full heap scan to locate the process object.
The results of the scan are cached and shared with the existing
findjsobject and findjsinstances commands so running any one of these
will setup the cache for the others.

This change also adds accessor functions for retrieving the values of
properties within JavaScript objects to llv8.cc/.h rather than just
formatting them as strings.