src: support V8 5.8 stack frame markers
On 64 bits systems, V8 stores SMIs (small ints) in the top 32 bits
of a 64 bits word.  Frame markers used to obey this convention but
as of V8 5.8, they are stored as 32 bits SMIs with the top half set
to zero.  Shift the raw value up to make it a normal SMI again.

Fixes test/frame-test.js with the latest node.js v8.0.0-pre, it was
no longer able to detect special JS stack frames like entry and exit
frames, argument adapter frames, etc.

PR-URL: https://github.com/nodejs/llnode/pull/84
Refs: https://github.com/nodejs/llnode/issues/81
Reviewed-By: Fedor Indutny <fedor@indutny.com>