configure: skip linking tools/gyp if it exists
Use the existing tools/gyp if it exists. It could be checked out
from source instead of being linked to the node-gyp installation.
Also corrects the names in the gyp linking routines.

Also tries to find the built library in out/Release/obj.target/
if it's not in out/Release/