src: introduce Printer class
Inspect methods (which are methods to visualize content from an object)
were tightly coupled with object definitions. In fact, the structure of
some objects is defined inside those Inspect methods, making it harder
to reuse code (for example, to return structured objects in our API
instead of strings). This commit is the beginning of a refactor effort
to decouple object visualization and object strucutres in order to
provide simpler code and a more useful API.

Most changes in this commit are moving ::Inspect methods from V8 Object
classes to a separate class Printer, and changing call sites to use
this new approach. Printer::Stringify is a single templated method with
specialization for each class it knows how to display.

PR-URL: https://github.com/nodejs/llnode/pull/246
Reviewed-By: Joyee Cheung <joyeec9h3@gmail.com>