src: fix JSError inspection with stringified stack
When the `stack` property of an Error object is accessed in V8, a
stringified version of the stack is generated, and then the "raw stack"
(with FP and arguments) is removed from memory. Our current inspection
code couldn't handle this because it was not checking if the raw stack
was a valid object. This commit makes that code more robust and thus
fixes inspection of error objects with stringified stacks.

PR-URL: https://github.com/nodejs/llnode/pull/291
Reviewed-By: Colin Ihrig <cjihrig@gmail.com>