src: fix findrefs for context locals
Previous code was too strict when hitting a context which it was unable
to read Locals information: it would stop looping through the contexts
even if we still had valid contexts in the queue. Changing it to loop
through all context and all Locals in a context should give us a more
reliable results.

Also, if a given object was found in a context Locals but we're unable
to read the Local name, we'll show it as ??? to indicate it's being
referenced but we couldn't read its name (in case of corrupted memory,
for example).

Fixes: https://github.com/nodejs/llnode/issues/278