src: treat embeded builtins as V8 functions
On Node.js v12 most builtins are compiled during build time and embeded
in the binary. Because of that, LLDB will think it knows how to handle
these frames (especially JavaScript frames). To correctly handle those
frames, we check the function name, if it starts with `Builtins_` we'll
handle it as a JIT function. This method should be safe since any C++
function coming from V8 or Node.js will be mangled, thus beginning with
`_Z...`. There might be a better way to handle this, but for now this
check should be enough.

PR-URL: https://github.com/nodejs/llnode/pull/301
Reviewed-By: Colin Ihrig <cjihrig@gmail.com>