test: make SessionOutput Promise tolerant
If a test case would use Promises for linesUntil, sometimes we would emit
lines events while there was no listener attached to it. This commit
changes SessionOutput to make it a little more resilient: now line
events will only be emitted when the SessionOutput is waiting. When wait
is called (via linesUntil, for example), it'll retrieve all lines (via
line events) on the buffer until it reaches a line matching the regexp.
If no lines match the regexp, we'll continue as before waiting for data
from lldb and buffering it.