tools: add script with commands to facilitate dev
Some tasks are very common when working on llnode, especially when
upgrading a V8 version. This new script is intended to make developers
life easier by simplifying tasks which usually require printfs on llnode
code or manually inspecting the memory of a process. Hopefully we'll add
more commands in the future.