Remove support for unboxed doubles
As of the mentioned upstream commit, the hack we use to detect the
possibility of unboxed doubles no longer works. Rather than trying to
create another hack, simply hardcode it to false. This requires us to
also remove support for EOL versions of Node.js which may still have
unboxed doubles.