scripts: installation without sparse-checkout
Old versions of Ubuntu (and likely other LTS distros) have `git`
versions which do not support `git-sparse-checkout`.
`git-sparse-checkout` is used as of #389.

Dynamically check if `git sparse-checkout` is supported, and if not then
fallback to cloning the whole repo. Cloning the whole repo is _slow_,
but at least it works.

I tested this patch on Ubuntu 18.04 without any lldb headers installed,
and it worked successfully.